Home work 3 - spring boot
