+++ hw3_testing/src/main/java/ru/otus/demen/Hw3TestingApplication.java
@@ -2,12 +2,16 @@ package ru.otus.demen;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.ConfigurableApplicationContext;
+import ru.otus.demen.service.TestingRunner;
 
 @SpringBootApplication
 public class Hw3TestingApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(Hw3TestingApplication.class, args);
+		ConfigurableApplicationContext context = SpringApplication.run(Hw3TestingApplication.class, args);
+		TestingRunner testingRunner = context.getBean(TestingRunner.class);
+		testingRunner.run();
 	}
 
 }
