+++ hw3_testing/src/main/resources/messages.properties
@@ -1,4 +1,3 @@
-# suppress inspection "InconsistentResourceBundle" for whole file
 input.name=Input your name:
 question=Question: {0}?
 output.name=Tested {0}
