+++ hw3_testing/src/test/java/ru/otus/demen/Hw3TestingApplicationTests.java
@@ -6,6 +6,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 @SpringBootTest
 class Hw3TestingApplicationTests {
 
+	@SuppressWarnings("EmptyMethod")
 	@Test
 	void contextLoads() {
 	}
