+++ hw3_testing/src/test/java/ru/otus/demen/service/TestResultCalculatorTest.java
@@ -1,36 +1,34 @@
 package ru.otus.demen.service;
 
 
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import ru.otus.demen.model.Question;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.ActiveProfiles;
 
 import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.*;
 
+@SpringBootTest
+@ActiveProfiles("test")
 class TestResultCalculatorTest {
+    @Autowired
     private TestResultCalculator testResultCalculator;
-    private List<Question> questions;
-
-    @BeforeEach
-    void setUp() {
-        testResultCalculator = new TestResultCalculator();
-        questions = List.of(new Question("Красный по английски", "red"),
-                new Question("Синий по английски", "blue"));
-    }
+    @Autowired
+    private TestProvider testProvider;
 
     @Test
     void checkTests() {
         List<String> answers = List.of("red", "black");
-        int successTestCounter = testResultCalculator.checkTests(questions, answers);
+        int successTestCounter = testResultCalculator.checkTests(testProvider.getTests(), answers);
         assertEquals(1, successTestCounter);
     }
 
     @Test
     void checkTests_caseInsensitive() {
         List<String> answers = List.of("Red", "black");
-        int successTestCounter = testResultCalculator.checkTests(questions, answers);
+        int successTestCounter = testResultCalculator.checkTests(testProvider.getTests(), answers);
         assertEquals(1, successTestCounter);
     }
 }
